+++ example/package.json
@@ -33,7 +33,7 @@
     "react-native-nitro-image": "^0.14.0",
     "react-native-nitro-modules": "^0.35.6",
     "react-native-vision-camera": "^5.0.9",
-    "react-native-vision-camera-ocr": "link:..",
+    "react-native-vision-camera-ocr-plus": "link:..",
     "react-native-vision-camera-worklets": "^5.0.9",
     "react-native-worklets": "^0.8.3"
   },
