+++ src/Camera.tsx
@@ -1,11 +1,11 @@
-import React, { forwardRef, type ForwardedRef, useMemo } from 'react'
-import * as VisionCameraModule from 'react-native-vision-camera'
+import React, { forwardRef, type ForwardedRef, useMemo } from 'react';
+import * as VisionCameraModule from 'react-native-vision-camera';
 import {
   createTextRecognitionPlugin,
   type TextRecognitionHandle,
-} from './scanText'
-import { createTranslatorPlugin, type TranslatorHandle } from './translateText'
-import { runOnJS } from 'react-native-worklets'
+} from './scanText';
+import { createTranslatorPlugin, type TranslatorHandle } from './translateText';
+import { runOnJS } from 'react-native-worklets';
 import type {
   CameraTypes,
   Text,
@@ -13,7 +13,7 @@ import type {
   ReadonlyFrameProcessor,
   TextRecognitionOptions,
   TranslatorOptions,
-} from './types'
+} from './types';
 
 /**
  * A drop-in replacement for VisionCamera's `<Camera />` that automatically
@@ -45,20 +45,20 @@ export const Camera = forwardRef(function Camera(
   props: CameraTypes,
   ref: ForwardedRef<any>
 ) {
-  const NativeCamera = (VisionCameraModule as any).Camera
+  const NativeCamera = (VisionCameraModule as any).Camera;
   const useFrameProcessor = (VisionCameraModule as any).useFrameProcessor as (
     processor: (frame: Frame) => void,
     deps?: ReadonlyArray<unknown>
-  ) => ReadonlyFrameProcessor
+  ) => ReadonlyFrameProcessor;
 
-  const { device, callback, options, mode, ...p } = props
+  const { device, callback, options, mode, ...p } = props;
 
   const recognizerHandle = useTextRecognition(
     mode === 'recognize' ? (options as TextRecognitionOptions) : undefined
-  )
+  );
   const translatorHandle = useTranslate(
     mode === 'translate' ? (options as TranslatorOptions) : undefined
-  )
+  );
 
   // Pull the raw Nitro HybridObjects out so the worklet compiler can capture
   // them directly. Nitro HybridObjects are JSI HostObjects — they are worklet-
@@ -66,56 +66,56 @@ export const Camera = forwardRef(function Camera(
   const recognizer =
     mode === 'translate'
       ? translatorHandle.recognizer
-      : recognizerHandle.recognizer
-  const translator = translatorHandle.translator
-  const fromLang = translatorHandle.from
-  const toLang = translatorHandle.to
+      : recognizerHandle.recognizer;
+  const translator = translatorHandle.translator;
+  const fromLang = translatorHandle.from;
+  const toLang = translatorHandle.to;
 
   // JS-thread handler for recognize mode
   const runData = useMemo(
     () => runOnJS((data: Text) => callback(data)),
     [callback]
-  )
+  );
 
   // JS-thread handler for translate mode
   const runTranslate = useMemo(
     () =>
       runOnJS((text: string) => {
-        if (!text) return
-        ;(translator as any)
+        if (!text) return;
+        (translator as any)
           .translate(text, fromLang, toLang)
-          .then((translated: string) => callback(translated))
+          .then((translated: string) => callback(translated));
       }),
     [translator, fromLang, toLang, callback]
-  )
+  );
 
   const processFrame = (frame: Frame): void => {
-    'worklet'
+    'worklet';
     // Call the Nitro HybridObject directly — no wrapper function involved.
     const nb = (frame as any).getNativeBuffer() as {
-      pointer: bigint
-      release: () => void
-    }
-    const orientation: string = (frame as any).orientation ?? 'up'
-    let ocrResult: Text | undefined | null
+      pointer: bigint;
+      release: () => void;
+    };
+    const orientation: string = (frame as any).orientation ?? 'up';
+    let ocrResult: Text | undefined | null;
     try {
       ocrResult = (recognizer as any).scanFrame(nb.pointer, orientation) as
         | Text
         | undefined
-        | null
+        | null;
     } finally {
-      nb.release()
+      nb.release();
     }
-    const result: Text = ocrResult ?? { resultText: '', blocks: [] }
+    const result: Text = ocrResult ?? { resultText: '', blocks: [] };
     if (mode === 'translate') {
       if (result.resultText) {
-        runTranslate(result.resultText)
+        runTranslate(result.resultText);
       }
     } else {
-      runData(result)
+      runData(result);
     }
-    ;(frame as any).dispose?.()
-  }
+    (frame as any).dispose?.();
+  };
 
   const frameProcessor = useFrameProcessor(processFrame, [
     recognizer,
@@ -125,7 +125,7 @@ export const Camera = forwardRef(function Camera(
     runData,
     runTranslate,
     mode,
-  ])
+  ]);
 
   return (
     <>
@@ -139,8 +139,8 @@ export const Camera = forwardRef(function Camera(
         />
       )}
     </>
-  )
-})
+  );
+});
 
 /**
  * React hook that creates and memoizes a text recognition handle.
@@ -148,12 +148,12 @@ export const Camera = forwardRef(function Camera(
 export function useTextRecognition(
   options?: TextRecognitionOptions
 ): TextRecognitionHandle {
-  return useMemo(() => createTextRecognitionPlugin(options), [options])
+  return useMemo(() => createTextRecognitionPlugin(options), [options]);
 }
 
 /**
  * React hook that creates and memoizes a translator handle.
  */
 export function useTranslate(options?: TranslatorOptions): TranslatorHandle {
-  return useMemo(() => createTranslatorPlugin(options), [options])
+  return useMemo(() => createTranslatorPlugin(options), [options]);
 }
