+++ src/PhotoRecognizer.ts
@@ -1,7 +1,7 @@
-import { NitroModules } from 'react-native-nitro-modules'
-import { Platform as RNPlatform } from 'react-native'
-import type { TextRecognizer } from './specs/TextRecognizer.nitro'
-import type { PhotoOptions, Text } from './types'
+import { NitroModules } from 'react-native-nitro-modules';
+import { Platform as RNPlatform } from 'react-native';
+import type { TextRecognizer } from './specs/TextRecognizer.nitro';
+import type { PhotoOptions, Text } from './types';
 
 /**
  * Recognize text in a still photo.
@@ -13,25 +13,25 @@ import type { PhotoOptions, Text } from './types'
  * console.log(result.resultText);
  */
 export async function PhotoRecognizer(options: PhotoOptions): Promise<Text> {
-  const { uri, orientation = 'portrait' } = options
+  const { uri, orientation = 'portrait' } = options;
 
   if (!uri) {
-    throw new Error("Can't resolve img uri")
+    throw new Error("Can't resolve img uri");
   }
 
   const recognizer =
-    NitroModules.createHybridObject<TextRecognizer>('TextRecognizer')
+    NitroModules.createHybridObject<TextRecognizer>('TextRecognizer');
 
   // Normalize URI per platform
-  let processUri = uri
+  let processUri = uri;
   if (RNPlatform.OS === 'ios') {
-    processUri = uri.replace('file://', '')
+    processUri = uri.replace('file://', '');
   } else {
-    const hasScheme = /^[a-zA-Z][a-zA-Z0-9+.-]*:\/\//.test(processUri)
+    const hasScheme = /^[a-zA-Z][a-zA-Z0-9+.-]*:\/\//.test(processUri);
     if (!hasScheme) {
-      processUri = `file://${processUri}`
+      processUri = `file://${processUri}`;
     }
   }
 
-  return recognizer.recognizePhoto(processUri, orientation) as Promise<Text>
+  return recognizer.recognizePhoto(processUri, orientation) as Promise<Text>;
 }
