+++ src/RemoveLanguageModel.ts
@@ -1,6 +1,6 @@
-import { NitroModules } from 'react-native-nitro-modules'
-import type { Translator } from './specs/Translator.nitro'
-import type { Languages } from './types'
+import { NitroModules } from 'react-native-nitro-modules';
+import type { Translator } from './specs/Translator.nitro';
+import type { Languages } from './types';
 
 /**
  * Remove a downloaded ML Kit translation language model.
@@ -10,6 +10,6 @@ import type { Languages } from './types'
  * @param code BCP-47 language code, e.g. 'fr', 'de', 'zh'
  */
 export async function RemoveLanguageModel(code: Languages): Promise<boolean> {
-  const translator = NitroModules.createHybridObject<Translator>('Translator')
-  return translator.removeLanguageModel(code)
+  const translator = NitroModules.createHybridObject<Translator>('Translator');
+  return translator.removeLanguageModel(code);
 }
