--- src/__tests__/PhotoRecognizer.test.d.ts
@@ -1,6 +1,0 @@
-declare const mockRecognizePhoto: jest.Mock<any, any, any>;
-declare const mockConfigure: jest.Mock<any, any, any>;
-declare const mockTextRecognizer: {
-    recognizePhoto: jest.Mock<any, any, any>;
-    configure: jest.Mock<any, any, any>;
-};
