--- src/__tests__/scanText.test.d.ts
@@ -1,1 +1,0 @@
-export {};
