--- src/__tests__/translateText.test.d.ts
@@ -1,1 +1,0 @@
-export {};
