+++ src/scanText.ts
@@ -1,16 +1,16 @@
-import { NitroModules } from 'react-native-nitro-modules'
-import type { TextRecognizer } from './specs/TextRecognizer.nitro'
+import { NitroModules } from 'react-native-nitro-modules';
+import type { TextRecognizer } from './specs/TextRecognizer.nitro';
 import type {
   TextRecognitionPlugin,
   TextRecognitionOptions,
   Text,
   Frame,
-} from './types'
+} from './types';
 
 export type TextRecognitionHandle = TextRecognitionPlugin & {
   /** The raw Nitro HybridObject - safe to capture directly in worklets. */
-  recognizer: TextRecognizer
-}
+  recognizer: TextRecognizer;
+};
 
 /**
  * Creates a plugin that synchronously recognizes text in a VisionCamera v5 frame.
@@ -34,31 +34,31 @@ export function createTextRecognitionPlugin(
         height: parseFloat(options.scanRegion.height),
       },
     }),
-  }
+  };
 
   const recognizer =
-    NitroModules.createHybridObject<TextRecognizer>('TextRecognizer')
-  recognizer.configure(config)
+    NitroModules.createHybridObject<TextRecognizer>('TextRecognizer');
+  recognizer.configure(config);
 
   return {
     recognizer,
     scanText: (frame: Frame): Text => {
-      'worklet'
+      'worklet';
       const nb = (frame as any).getNativeBuffer() as {
-        pointer: bigint
-        release: () => void
-      }
-      const orientation: string = (frame as any).orientation ?? 'up'
-      let result: Text | undefined | null
+        pointer: bigint;
+        release: () => void;
+      };
+      const orientation: string = (frame as any).orientation ?? 'up';
+      let result: Text | undefined | null;
       try {
         result = (recognizer as any).scanFrame(nb.pointer, orientation) as
           | Text
           | undefined
-          | null
+          | null;
       } finally {
-        nb.release()
+        nb.release();
       }
-      return result ?? ({ resultText: '', blocks: [] } as unknown as Text)
+      return result ?? ({ resultText: '', blocks: [] } as unknown as Text);
     },
-  }
+  };
 }
