+++ src/specs/Translator.nitro.d.ts
@@ -4,16 +4,16 @@ import type { HybridObject } from 'react-native-nitro-modules';
  * Handles downloading language models, translating text, and removing models.
  */
 export interface Translator extends HybridObject<{
-    ios: 'swift';
-    android: 'kotlin';
+  ios: 'swift';
+  android: 'kotlin';
 }> {
-    /**
-     * Translate text from one language to another.
-     * Downloads the required model on first use if not already present.
-     */
-    translate(text: string, from: string, to: string): Promise<string>;
-    /**
-     * Remove a downloaded translation language model to free device storage.
-     */
-    removeLanguageModel(languageCode: string): Promise<boolean>;
+  /**
+   * Translate text from one language to another.
+   * Downloads the required model on first use if not already present.
+   */
+  translate(text: string, from: string, to: string): Promise<string>;
+  /**
+   * Remove a downloaded translation language model to free device storage.
+   */
+  removeLanguageModel(languageCode: string): Promise<boolean>;
 }
