+++ src/specs/VisionCameraOcrPlus.nitro.d.ts
@@ -1,2 +1,12 @@
-export type { TextRecognizer, RecognizedText, TextBlock, TextLine, TextElement, BoundingFrame, CornerPoint, ScanRegion, TextRecognitionConfig, } from './TextRecognizer.nitro';
+export type {
+  TextRecognizer,
+  RecognizedText,
+  TextBlock,
+  TextLine,
+  TextElement,
+  BoundingFrame,
+  CornerPoint,
+  ScanRegion,
+  TextRecognitionConfig,
+} from './TextRecognizer.nitro';
 export type { Translator } from './Translator.nitro';
