+++ tsconfig.json
@@ -1,20 +1,19 @@
 {
-  "include": [
-    "./src"
-  ],
   "compilerOptions": {
-    "composite": true,
-    "outDir": "lib",
-    "rootDir": "./src",
+    "rootDir": ".",
+    "paths": {
+      "react-native-vision-camera-ocr-plus": ["./src/index"]
+    },
     "allowUnreachableCode": false,
     "allowUnusedLabels": false,
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
     "jsx": "react",
     "lib": ["esnext"],
+    "types": ["jest"],
     "module": "esnext",
     "moduleResolution": "node",
-    "noEmit": false,
+    "ignoreDeprecations": "6.0",
     "noFallthroughCasesInSwitch": true,
     "noImplicitReturns": true,
     "noImplicitUseStrict": false,
@@ -27,5 +26,6 @@
     "strict": true,
     "target": "esnext",
     "verbatimModuleSyntax": true
-  }
+  },
+  "exclude": ["lib"]
 }
