--- v4_files/babel.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  presets: ['module:@react-native/babel-preset'],
-};
