--- v4_files/package.json
@@ -1,154 +1,0 @@
-{
-  "name": "react-native-vision-camera-ocr-plus",
-  "version": "1.2.9",
-  "description": "React Native Vision Camera plugin for on-device text recognition (OCR) and translation using ML Kit. Maintained fork of react-native-vision-camera-text-recognition",
-  "main": "lib/commonjs/index",
-  "module": "lib/module/index",
-  "types": "lib/typescript/src/index.d.ts",
-  "react-native": "src/index",
-  "source": "src/index",
-  "files": [
-    "src",
-    "lib",
-    "android",
-    "ios",
-    "cpp",
-    "*.podspec",
-    "!ios/build",
-    "!android/build",
-    "!android/gradle",
-    "!android/gradlew",
-    "!android/gradlew.bat",
-    "!android/local.properties",
-    "!**/__tests__",
-    "!**/__fixtures__",
-    "!**/__mocks__",
-    "!**/.*"
-  ],
-  "scripts": {
-    "test": "jest",
-    "test:coverage": "jest --coverage",
-    "typecheck": "tsc --noEmit",
-    "lint": "eslint \"**/*.{js,ts,tsx}\"",
-    "clean": "del-cli android/build example/android/build example/android/app/build example/ios/build lib",
-    "prepare": "bob build",
-    "release": "standard-version"
-  },
-  "keywords": [
-    "react-native",
-    "ios",
-    "android",
-    "camera",
-    "mlkit",
-    "ML Kit",
-    "react-native-vision-camera",
-    "vision-camera",
-    "text-recognition",
-    "text",
-    "ocr",
-    "text",
-    "text-detector",
-    "text-recognizer",
-    "mlkit",
-    "google-mlkit",
-    "machine-learning",
-    "image-processing",
-    "react-native-camera",
-    "react-native-ocr",
-    "vision-camera-ocr",
-    "react-native-vision-camera-text-recognition"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/jamenamcinteer/react-native-vision-camera-ocr-plus"
-  },
-  "author": "Jamena McInteer <hello@jamena.dev> (https://github.com/jamenamcinteer)",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/jamenamcinteer/react-native-vision-camera-ocr-plus/issues"
-  },
-  "homepage": "https://github.com/jamenamcinteer/react-native-vision-camera-ocr-plus#readme",
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org/"
-  },
-  "devDependencies": {
-    "@commitlint/config-conventional": "^20.5.0",
-    "@eslint/eslintrc": "^3.3.5",
-    "@eslint/js": "^10.0.1",
-    "@react-native/babel-preset": "^0.85.1",
-    "@react-native/eslint-config": "^0.85.1",
-    "@react-native/jest-preset": "^0.85.1",
-    "@types/jest": "^30.0.0",
-    "@types/react": "^19.2.14",
-    "commitlint": "^20.5.0",
-    "del-cli": "^7.0.0",
-    "eslint": "^9.39.4",
-    "eslint-config-prettier": "^10.1.8",
-    "eslint-plugin-ft-flow": "^3.0.11",
-    "eslint-plugin-prettier": "^5.5.5",
-    "jest": "^30.3.0",
-    "prettier": "^3.8.2",
-    "react": "19.2.5",
-    "react-native": "^0.85.1",
-    "react-native-builder-bob": "^0.41.0",
-    "react-native-vision-camera": "^5.0.9",
-    "react-native-vision-camera-worklets": "^5.0.9",
-    "react-native-worklets": "^0.8.3",
-    "react-native-worklets-core": "^1.6.3",
-    "standard-version": "^9.5.0",
-    "typescript": "^6.0.2"
-  },
-  "resolutions": {
-    "@types/react": "^19.1.0"
-  },
-  "peerDependencies": {
-    "react": "*"
-  },
-  "jest": {
-    "preset": "@react-native/jest-preset",
-    "modulePathIgnorePatterns": [
-      "<rootDir>/example/node_modules",
-      "<rootDir>/lib/"
-    ],
-    "testMatch": [
-      "<rootDir>/src/**/__tests__/**/*.{js,jsx,ts,tsx}",
-      "<rootDir>/src/**/*.(test|spec).{js,jsx,ts,tsx}"
-    ],
-    "collectCoverageFrom": [
-      "src/**/*.{js,jsx,ts,tsx}",
-      "!src/**/*.d.ts",
-      "!src/__tests__/**/*"
-    ],
-    "coverageDirectory": "coverage",
-    "coverageReporters": [
-      "text",
-      "lcov",
-      "html"
-    ],
-    "testEnvironment": "node",
-    "transformIgnorePatterns": [
-      "node_modules/(?!(react-native|@react-native|react-native-vision-camera|react-native-worklets-core)/)"
-    ]
-  },
-  "commitlint": {
-    "extends": [
-      "@commitlint/config-conventional"
-    ]
-  },
-  "react-native-builder-bob": {
-    "source": "src",
-    "output": "lib",
-    "targets": [
-      "commonjs",
-      "module",
-      [
-        "typescript",
-        {
-          "project": "tsconfig.build.json"
-        }
-      ]
-    ]
-  },
-  "packageManager": "yarn@1.22.21+sha1.1959a18351b811cdeedbd484a8f86c3cc3bbaf72",
-  "dependencies": {}
-}
